add integration test for embedded resource file
There will eventually need to be tests around ensuring that upgrading
versions adds additional content to the config.ini but for now this test
is all that is needed for resource files. This test is simply a
safeguard to prevent against developer error (renames/deletes/etc). It
serves no other practical purpose.